api/requirements.txt
aniso8601==4.0.1
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.7
Flask-JWT==0.3.2
Flask-PyMongo==2.2.0
Flask-RESTful==0.3.7
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
PyJWT==1.4.2
pymongo==3.7.2
pytz==2018.7
simplejson==3.16.0
six==1.12.0
Werkzeug==0.14.1
